# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mind Merge API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/task/task_tree": {
      "get": {
        "summary": "Get all tasks for a user and organization",
        "description": "Get all tasks for a user with given id and organization with given id",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "user_id",
            "description": "The id of the user",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "organization_id",
            "description": "The id of the organization",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Token",
            "description": "The jwt (json web token) of the user",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully returns the task tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/task/{task_id}": {
      "get": {
        "summary": "Get a task",
        "description": "Get a task with given id and valid organization id",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "task_id",
            "description": "The id of the task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "organization_id",
            "description": "The id of the organization",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Token",
            "description": "The jwt (json web token) of the user",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully returns the Task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "Task"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Delete a task with given id and valid organization id",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "task_id",
            "description": "The id of the task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "organization_id",
            "description": "The id of the organization",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Token",
            "description": "The jwt (json web token) of the user",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully removes the task"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/task/{task_id}/name/{new_name}": {
      "put": {
        "summary": "Update a task name",
        "description": "Update a task name with given id and valid organization id",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "task_id",
            "description": "The id of the task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "new_name",
            "description": "The new name of the task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organization_id",
            "description": "The id of the organization",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Token",
            "description": "The jwt (json web token) of the user",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updates the task name"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/task/{task_id}/notes/{notes_id}": {
      "put": {
        "summary": "Update a task note",
        "description": "Update a task note with given id and valid organization id",
        "tags": [
          "Tasks"
        ],
        "requestBody": {
          "description": "The new notes for the task",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "task_id",
            "description": "The id of the task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "notes_id",
            "description": "The id of the note",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organization_id",
            "description": "The id of the organization",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Token",
            "description": "The jwt (json web token) of the user",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updates the task name"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/task/{task_id}/notes/": {
      "post": {
        "summary": "Create a new task note",
        "description": "Create a new task note with given data",
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "task_id",
            "description": "The id of the task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Token",
            "description": "The jwt (json web token) of the user",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "name": "Task",
          "description": "The Task data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully creates the task note"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}